Support use of a map.mission.txt.rename to rename obsid values from map.mission.txt to match the obscore value (MAST, see issue #22)
This is really a hack. Whilst it allows the papers to include the data
references, the links to the quick-look EUVE products are not correct
(or at least not for yzcmi__9412210116N) since they seem to need the
"old" obs id names (as used by the bibliography list rather than in the
obscore table).
(cherry picked from commit c63bac9aff39008aff88917fba07b6cf85ad1b13)